Enable Debug builds on GCC 6.
GCC 6 issues a warning when a `throw` expression would immediately result
in a call to `terminate`.  Because Debug builds are built with `-Werror`,
this means that the build fails.

So, I disable this warning (`-Wno-terminate`).  Of course, it might be
preferable to change the code to not trigger this warning in the first
place.