Add .gitignore files for in-tree builds
I know the "recommended" process is to do out-of-tree builds in another
directory, but in-tree builds also work.  When building from git, this
means that it is easy to accidentally add build files to the git repo.

I believe that the ignore lists are complete on "unix" builds, though I
believe that on Windows the list is still incomplete.